pyproject.toml
[project]
name = "llm-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.31.0",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.12",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.58",
    "langchain-ollama>=0.3.2",
    "python-dotenv>=1.1.0",
]
